Split CopyFromMssToPg tests
